+++ ARACA_REPORT_20250731-020453.md
@@ -1,0 +1,39 @@
+# ARACA Report (2025-07-31T02:04:53Z)
+
+## Objectives and Constraints
+- Perform structural improvements without replacing Eleventy, Nunjucks or Tailwind.
+- Reduce redundancy and clarify boundaries across the codebase.
+- Provide deterministic validation via tests.
+
+## Inventory of Changes
+- `lib/filters.js` – standalone utilities for date and collection formatting.
+- `lib/markdown-extensions.js` – modular markdown-it enhancements for footnotes and embeds.
+- `lib/plugins.js` – centralised Eleventy plugin configuration.
+- `.eleventy.js` – simplified by consuming the above modules.
+- `README.md` – updated snippet showing new configuration pattern.
+- `test/filters.test.js` – minimal Node.js test suite exercising filter behaviour.
+- `package.json` – `test` script uses Node's built‑in runner.
+
+## Refactor Summary
+- **Modularisation**: extracted large inline sections from `.eleventy.js` into focused modules for plugins, filters and markdown extensions.
+- **Configuration hygiene**: consolidated plugin setup via `getPlugins()` and registered filters programmatically.
+- **Type strengthening**: added JSDoc comments describing parameters and return values for exported utilities.
+- **Testing**: introduced `node:test` based checks ensuring filters behave deterministically.
+
+## Validation Results
+- `npm install` completed successfully.
+- `npm test` – all 4 tests pass.
+
+## Performance / Footprint
+- No runtime dependencies added; tests use existing packages only.
+
+## Citations
+- Filters module implementation【F:lib/filters.js†L1-L65】
+- Markdown extension module implementation【F:lib/markdown-extensions.js†L1-L40】
+- Plugin configuration module implementation【F:lib/plugins.js†L1-L36】
+- Updated Eleventy configuration utilising modules【F:.eleventy.js†L1-L40】
+- README excerpt showing new configuration usage【F:README.md†L60-L72】
+
+## Next Steps
+- Expand test coverage for remaining utilities and shortcodes.
+- Consider linting (ESLint) and formatting to ensure consistent style across modules.
